wip gpu update

+++ Technolize/World/Block/BlockInfo.cs
@@ -3,9 +3,9 @@ using Raylib_cs;
 using Technolize.World.Tag;
 namespace Technolize.World.Particle;
 
-public sealed class BlockInfo(long id, Vector2 size, MatterState matterState, Color color)
+public sealed class BlockInfo(int id, Vector2 size, MatterState matterState, Color color)
 {
-    public long Id { get; } = id;
+    public int Id { get; } = id;
     public Vector2 Size { get; } = size;
     public MatterState MatterState { get; } = matterState;
     public Color Color { get; } = color;
